app/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.4
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-formtools==2.4
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
django-otp==1.1.3
django-phonenumber-field==6.4.0
django-two-factor-auth==1.14.0
djangorestframework==3.12.4
drf-spectacular==0.15.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.0
jwcrypto==1.4.2
Markdown==3.4.1
oauthlib==3.2.2
packaging==21.3
phonenumbers==8.12.57
Pillow==9.3.0
psycopg2==2.9.4
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
qrcode==7.3.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
celery==5.2.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-rosetta==0.9.8